requirements.txt
torch
transformers
bitsandbytes
onnx
onnxruntime
datasets
scikit-learn
matplotlib
tqdm
Pillow
ultralytics
opencv-python
